Update README.md
#RYTHMIC TUNES

## 🆔 Team Details  
- *Team ID:* SWTID1741163107154439
- *Team Leader:*Rakesh K
- *Team Members:*  
  - Murali T  
  - Vetri S 
  - Mohanasundharam S 

## 📌 Project Description  
A rhythmic tune is characterized by a structured sequence of sounds and silences that create a pattern over time. It serves as the heartbeat of a melody, guiding its pace and emotional expression. Focusing on rhythm can enhance a tune's drive, energy, and memorability, making it a crucial element in music composition. Overall, rhythm shapes how music is experienced, influencing its dynamics and flow

## 🛠 Technologies Used  
- *Frontend:* React.js  
- *Backend:* Node.js 
- *Database:* MongoDB / Firebase  
- *API Integration:* Spotify API/Apple Music API

## 🚀 How to Run  
1. Clone the repository:  
   ```bash
   https://github.com/Murali123098/music-app/edit/main/code/README.md